Abstract donation progress bar
